Feature.
Updated use of the Role enumeration.
Moved the swagger files to a common directory.
Moved the logger configuration files to a debug module.
Removed the test jar and test Java documentation artifacts.
Disabled the JavaDoc for the model module.
Added the version number to the index and introduction pages.
Updated the CSS for color and the standard mule shell theme.
Added identifier to MuleShell.
Added parameter for setting/configuration of caches.